# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IChing MVP API
  version: 1.0.0
  description: MVP backend - user, bazi, order, admin
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/order/create:
    post:
      tags: [order]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productType]
              properties:
                productType: { type: integer, description: '1=deep_report 2=membership' }
                paipanId: { type: integer }
      responses:
        '200': { description: Order created }
  /api/order/pay:
    post:
      tags: [order]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderNo]
              properties:
                orderNo: { type: string }
      responses:
        '200': { description: Mock payment success }
  /api/order/{orderNo}:
    get:
      tags: [order]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderNo
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Order status }
...